+++ src/e2e-helper/e2e-command-helper.ts
@@ -824,12 +824,20 @@ export default class CommandHelper {
     });
     return this.runCmd(`bit _snap '${JSON.stringify(data)}' ${options}`, cwd);
   }
-  snapFork(data: Record<string, any>, options = '') {
-    return this.snapFromScope(this.scopes.localPath, data, options);
-  }
   snapFromScopeParsed(cwd: string, data: Record<string, any>, options = '') {
     const output = this.snapFromScope(cwd, data, `${options} --json`);
     return JSON.parse(output);
+  }
+  apply(data: Record<string, any>, options = '') {
+    data.forEach((dataItem) => {
+      if (!dataItem.files) return;
+      dataItem.files.forEach((file) => {
+        if (file.content) {
+          file.content = Buffer.from(file.content).toString('base64');
+        }
+      });
+    });
+    return this.runCmd(`bit apply '${JSON.stringify(data)}' ${options}`);
   }
   diff(id = '') {
     const output = this.runCmd(`bit diff ${id}`);
